+++ lib/avromatic/model/types/array_type.rb
@@ -57,8 +57,9 @@ module Avromatic
             end
 
             if missing_attributes.present?
-              raise Avromatic::Model::ValidationError.new('Array cannot be serialized because the ' \
-                "following attributes are nil: #{missing_attributes.map(&:to_s).join(', ')}", missing_attributes)
+              message = 'Array cannot be serialized because the following attributes are nil: ' \
+                  "#{missing_attributes.map(&:to_s).join(', ')}"
+              raise Avromatic::Model::ValidationError.new(message, missing_attributes)
             else
               avro_hash
             end